% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{deploy_MAST21V3_2024_v2}
\alias{deploy_MAST21V3_2024_v2}
\title{Run the MAST21 2024 v2 protocol}
\usage{
deploy_MAST21V3_2024_v2(
  app_name = paste("UPEI ", format(Sys.Date(), "\%Y"), " Study"),
  language,
  musicassessr_aws = FALSE,
  data_collection_method = "audio",
  get_p_id = FALSE,
  final_qualtrics_url =
    "https://upeiairs.qualtrics.com/jfe/form/SV_5vDAjJhxLqZw7Km?participant=",
  opening_and_final_image =
    "https://img.freepik.com/free-vector/headphone-concept-illustration_114360-2132.jpg?w=826&t=st=1708543460~exp=1708544060~hmac=7a9cba8f7104f82422a55f30f5120598e4ce64dd222a1247c8da79825dddacb6"
)
}
\arguments{
\item{opening_and_final_image}{}
}
\description{
Run the MAST21 2024 v2 protocol
}
