% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAST_helpers.R
\name{get_dob_page}
\alias{get_dob_page}
\title{deploy the MAST21 as wavs}
\usage{
get_dob_page(text = "When is your date of birth?")
}
\arguments{
\item{musicassessr_aws}{}
}
\description{
deploy the MAST21 as wavs
}
